ApiWrapper and Callback implementation
Implementing the logic required to fix the previous confusing code.
"And just look at the code once again. Is it as pretty as before? No! It’s awful. Now it has much more code noise, more curly braces, but the logic is the same.
If we look closer to our Callback interfaces we can spot the common pattern:

All of them have a method that delivers the result (onCutestCatSaved, onCatListReceived, onCatStored)
Most of them (all in our case) have a method to handle errors that happen during operation (onError, onQueryFailed, onStoreFailed)
So we can create a generic callback to replace them all. But as we can not change api call signatures we would have to create wrapped calls."